move some code from the erlang package to the fetch package
