+++ gazelle/BUILD.bazel
@@ -16,9 +16,7 @@ go_library(
         "erlang_app.go",
         "fix.go",
         "generate.go",
-        "github_com.go",
         "hex_metadata_parser.go",
-        "hex_pm.go",
         "kinds.go",
         "language.go",
         "moduleindex.go",
