+++ gazelle/fetch/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "fetch",
     srcs = [
+        "fetch.go",
         "github.go",
         "hex_pm.go",
         "tar.go",
