+++ gazelle/fetch/github.go
@@ -2,6 +2,8 @@ package fetch
 
 import (
 	"fmt"
+	"net/url"
+	"path"
 	"regexp"
 )
 
@@ -24,4 +26,14 @@ func ParseGithubImportArg(imp string) (name, version, owner, repo, ref string, e
 		err = fmt.Errorf("not a valid import string: %s", imp)
 	}
 	return
+}
+
+func DownloadRef(owner, repo, ref, filepath string) error {
+	url := url.URL{
+		Scheme: "https",
+		Host:   "github.com",
+		Path:   path.Join(owner, repo, "archive", ref+".tar.gz"),
+	}
+
+	return Download(&url, filepath)
 }
