+++ gazelle/util.go
@@ -2,10 +2,6 @@ package erlang
 
 import (
 	"fmt"
-	"io"
-	"net/http"
-	"net/url"
-	"os"
 
 	"github.com/bazelbuild/bazel-gazelle/config"
 )
@@ -58,29 +54,4 @@ func Log(c *config.Config, a ...interface{}) (n int, err error) {
 		return fmt.Println(a...)
 	}
 	return 0, nil
-}
-
-func Download(url *url.URL, filepath string) error {
-	resp, err := http.Get(url.String())
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("bad status: %s", resp.Status)
-	}
-
-	out, err := os.Create(filepath)
-	if err != nil {
-		return err
-	}
-	defer out.Close()
-
-	_, err = io.Copy(out, resp.Body)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
